requirements.txt
openai
pyttsx3
tomli
fastapi
pydub
python-multipart
uvicorn
gtts
git+https://github.com/openai/whisper.git
azure-cognitiveservices-speech